Starting proc macros to avoid pointers
This is a checkpoint of work in progress. The goal is to move to
accessor functions and away from direct use of device pointers, as the
latter are not portable. Towards this goal, we will code-generate
headers using Rust procedural macros. These same macros will also help
automate the encoding process.