Cleanup
Rename piet-metal-derive to piet-gpu-derive, now that it handles both
HLSL and Metal. Remove a bunch of old code. Change generators to accept
shading language argument.

This is a bit of a snapshot to restore the piet-dx12 functionality, but
isn't fully at the point where it can replace the accessors for
piet-metal.